student-insight-system/migrations/versions/add_department_to_skills.py
"""add department to skills

Revision ID: a2b3c4d5e6f7
Revises: f1a2b3c4d5e6
Create Date: 2026-03-03

"""
from alembic import op
import sqlalchemy as sa

# revision identifiers, used by Alembic.
revision = 'a2b3c4d5e6f7'
down_revision = 'f1a2b3c4d5e6'
branch_labels = None
depends_on = None


def upgrade():
    op.add_column('skills', sa.Column('department', sa.String(100), nullable=True))


def downgrade():
    op.drop_column('skills', 'department')
