student-insight-system/requirements/base.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flask-migrate==4.1.0
psycopg2-binary==2.9.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy>=2.0.0
PyJWT==2.10.1
Werkzeug==3.1.5
python-dotenv==1.0.0
Flask-SQLAlchemy
scikit-learn
google-generativeai
